CIV-1928 Update dependencies and run new lint rules
