+++ src/UCATemplateValue.js
@@ -1,0 +1,13 @@
+class UCATemplateValue {
+  constructor(identifier, type, propertyName, required, version) {
+    this.name = identifier;
+    this.meta = {
+      required,
+      propertyName,
+      type,
+      version,
+    };
+  }
+}
+
+module.exports = { UCATemplateValue };
